backend/data-analyser/requirements.txt
matplotlib==3.10.1
nats-py==2.10.0
pandas==2.2.3
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.1.0